code/Sample GCF Application/requirements.txt
google-cloud-storage
google-cloud-pubsub
google-cloud-trace==0.19.0
opencensus==0.1.8